+++ print_digits.c
@@ -6,7 +6,7 @@
 /*   By: mbuchet <mbuchet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/19 15:28:26 by mbuchet           #+#    #+#             */
-/*   Updated: 2026/04/20 18:58:01 by mbuchet          ###   ########.fr       */
+/*   Updated: 2026/04/20 19:27:07 by mbuchet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void	print_number_unsigned(unsigned int number, size_t *str_lenght)
 	print_char((n_long % 10) + '0', str_lenght);
 }
 
-void	print_hexadecimal(unsigned int number, int is_uppercase, size_t *str_lenght)
+void	print_hexadecimal(unsigned long number, int is_uppercase, size_t *str_lenght)
 {
 	unsigned long	n_long;
 	char		*str;
@@ -51,3 +51,4 @@ void	print_hexadecimal(unsigned int number, int is_uppercase, size_t *str_lenght
 		print_hexadecimal(n_long / 16, is_uppercase, str_lenght);
 	print_char(str[n_long % 16], str_lenght);
 }
+
